+++ scripts/creatures/pack-tactics.js
@@ -7,25 +7,13 @@ const optionName = "Pack Tactics";
 try {
 	if (args[0].macroPass === "preAttackRoll") {
 		const targetToken = workflow.targets.first();
-		const creatures = MidiQOL.findNearby(CONST.TOKEN_DISPOSITIONS.HOSTILE, targetToken, 5);
-		const validCreatures = arrayRemove(creatures, token);
+		const nearbyAllies = MidiQOL.findNearby(CONST.TOKEN_DISPOSITIONS.HOSTILE, targetToken, 5);
 
-		if (validCreatures.length > 0) {
-			const eligibleAllies =
-		}
-		
-		if (validCreatures.length === 0 || validCreatures.filter(t=>t.actor.effects.find(i=>i.name === "Incapacitated")).length === validCreatures.length)
-			return;
-		else
+		if (nearbyAllies.length > 0) {
 			workflow.advantage = true;
+		}
 	}
 
 } catch (err) {
     console.error(`${optionName}: ${version}`, err);
-}
-
-function arrayRemove(arr, value) {
-    return arr.filter(function(ele){
-        return ele != value;
-    });
 }